Make Login Scenario Split
Thre was several scenario in user login scenario and we want to make sure there are no exception can happen. But to do that only with existing logical state was impossible. So we decide to split and make new logical state loading, which is you can see in other app during loading page such as kakao talk.